frontend/auditpulse_mvp/requirements.txt
# Core Dependencies
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
alembic==1.12.0
pydantic==2.3.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
psycopg2-binary==2.9.7
asyncpg==0.28.0
aiosqlite==0.19.0
tenacity==8.2.3
httpx==0.24.1
aiohttp==3.8.5

# API Auth & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
cryptography==41.0.3
authlib==1.2.1

# Data Connectors
plaid-python==16.0.0
intuit-oauth==1.2.5
quickbooks-python==0.9.8

# ML & Analytics
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.25.2
joblib==1.3.2
statsmodels==0.14.0

# LLM Integration
openai==0.28.0
tiktoken==0.4.0
langchain==0.0.289
llama-index==0.8.19
transformers==4.33.1

# Dashboard & Reporting
streamlit==1.27.0
plotly==5.16.1
matplotlib==3.7.3
reportlab==4.0.4
jinja2==3.1.2

# Task Scheduling & Queue
apscheduler==3.10.4
celery==5.3.4
redis==5.0.0
flower==2.0.1

# Notifications
slackclient==2.9.4
sendgrid==6.10.0
twilio==8.8.0

# Monitoring & Metrics
prometheus-client==0.17.1
prometheus-fastapi-instrumentator==6.1.0
structlog==23.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1
requests-mock==1.11.0
faker==19.6.1

# Dev Tools
black==23.7.0
flake8==6.1.0
mypy==1.5.1
bandit==1.7.5
pre-commit==3.4.0
safety==2.3.5
isort==5.12.0
types-requests==2.31.0.2
types-python-jose==3.3.4.8
types-redis==4.6.0.3 